abstract out the api
